CY_LLM_Backend/gateway_lite/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
grpcio==1.70.0
protobuf==5.29.3
pydantic==2.10.6
